requirements.txt
Flask==3.0.3
pymongo[srv]==4.10.1
dnspython==2.7.0
python-dotenv==1.0.1
